+++ source/db_extensions/extra/constraints.d
@@ -1,28 +1,43 @@
+/**
+User-defined attributes that can be used with the KeyedItem mixin.
+ */
 module db_extensions.extra.constraints;
 
 import std.functional : unaryFun;
 
 /**
-User-defined attribute that can be used with KeyedItem. KeyedItem
-will create a struct made up of all of the properties marked with
-@PrimaryKeyColumn which can be used with KeyedCollection as
-keys in an associative array by default.
- */
-alias PrimaryKeyColumn = UniqueConstraintColumn!("PrimaryKey");
-/**
-User-defined attribute that can be used with KeyedItem. KeyedItem
-will create a struct with name defined in the compile-time argument.
+KeyedItem will create a struct with *name* defined in the compile-time argument.
 For example a property marked with @UniqueColumn!("uc_Person") will
 be part of the struct uc_Person.
  */
 struct UniqueConstraintColumn(string name_)
 {
-    /// The name of the constraint which is the structs name.
+/**
+The name of the constraint which is the structs name.
+ */
     enum name = name_;
 }
 
-struct CheckConstraint(alias check_)
+/**
+An alias for the primary key column.
+ */
+alias PrimaryKeyColumn = UniqueConstraintColumn!("PrimaryKey");
+
+
+/**
+KeyedItem.checkConstraints will check all of the members marked
+with this attribute and use the check given.
+ */
+struct CheckConstraint(alias check_, string name_ = "")
     if (is(typeof(unaryFun!check_)))
 {
+/**
+The function that returns a boolean.
+ */
     alias check = unaryFun!check_;
+/**
+Name used in the error message if the function returns false. This may
+help narrow down which constraint failed.
+ */
+    enum name = name_;
 }